+++ headjack/__init__.py
@@ -1,3 +1,3 @@
-__all__ = ["agents", "api", "models", "tools"]
+__all__ = ["agents", "api", "models"]
 
-from headjack import agents, api, models, tools
+from headjack import agents, api, models
