+++ headjack/agents/knowledge_search.py
@@ -1,9 +1,10 @@
 import logging
 
 import lmql
+from headjack.agents.registry import register_agent_function
 
 from headjack.config import get_settings
-from headjack.utils.general import fetch
+from headjack.utils import fetch
 
 _logger = logging.getLogger("uvicorn")
 
@@ -19,7 +20,7 @@ async def search_for_knowledge(q):
         _logger.info("Error while attempting to reach headjack search " f"service knowledge collection: {str(e)}")
         return "No results"
 
-
+@register_agent_function("This is a general knowledge search. Provided a query, this will give a summary of information from the knowledge base.", "knowledge_search_agent")
 @lmql.query
 async def knowledge_search_agent(question: str):
     """
