+++ headjack/agents/metric_search.py
@@ -3,8 +3,8 @@ import logging
 import lmql
 
 from headjack.config import get_settings
-from headjack.utils.general import fetch
-
+from headjack.utils import fetch
+from headjack.agents.registry import register_agent_function
 _logger = logging.getLogger("uvicorn")
 
 
@@ -19,7 +19,7 @@ async def search_for_metrics(q, n: int = 5):
         _logger.info("Error while attempting to reach the headjack " f"search service metrics collection: {str(e)}")
         return "No results"
 
-
+@register_agent_function("""This function takes a query to search for some metric which is a value that can be calculated such as average, total, etc.""", "metric_search_agent")
 @lmql.query
 async def metric_search_agent(question: str):
     '''
