+++ headjack/utils/__init__.py
@@ -1,0 +1,1 @@
+from .fetch import fetch
